+++ src/renderer/src/i18n/locales/en-us.json
@@ -851,7 +851,8 @@
         "check": "Check",
         "search_provider": "Search service provider",
         "search_provider_placeholder": "Choose a search service provider.",
-        "no_provider_selected": "Please select a search service provider before checking."
+        "no_provider_selected": "Please select a search service provider before checking.",
+        "check_failed": "Verification failed"
       }
     },
     "translate": {
