+++ src/renderer/src/i18n/locales/ru-ru.json
@@ -851,7 +851,8 @@
         "check": "проверка",
         "search_provider": "поиск сервисного провайдера",
         "search_provider_placeholder": "Выберите поставщика поисковых услуг",
-        "no_provider_selected": "Пожалуйста, выберите поставщика поисковых услуг, затем проверьте."
+        "no_provider_selected": "Пожалуйста, выберите поставщика поисковых услуг, затем проверьте.",
+        "check_failed": "Проверка не прошла"
       }
     },
     "translate": {
