+++ src/renderer/src/pages/settings/WebSearchSettings/index.tsx
@@ -1,158 +1,107 @@
-import { SettingOutlined } from '@ant-design/icons'
-import { DragDropContext, Draggable, Droppable, DropResult } from '@hello-pangea/dnd'
-import Scrollbar from '@renderer/components/Scrollbar'
-import { getWebSearchProviderLogo } from '@renderer/config/webSearchProviders'
-import { useWebSearchProviders } from '@renderer/hooks/useWebSearchProviders'
+import { CheckOutlined, InfoCircleOutlined, LoadingOutlined } from '@ant-design/icons'
+import { useTheme } from '@renderer/context/ThemeProvider'
+import { useDefaultWebSearchProvider, useWebSearchProviders } from '@renderer/hooks/useWebSearchProviders'
+import WebSearchService from '@renderer/services/WebSearchService'
 import { WebSearchProvider } from '@renderer/types'
-import { droppableReorder } from '@renderer/utils'
-import { Avatar, Tag } from 'antd'
+import { Button, Select } from 'antd'
 import { FC, useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import styled from 'styled-components'
 
+import { SettingContainer, SettingDivider, SettingGroup, SettingRow, SettingRowTitle, SettingTitle } from '..'
 import BasicSettings from './BasicSettings'
+import BlacklistSettings from './BlacklistSettings'
 import WebSearchProviderSetting from './WebSearchProviderSetting'
 
 const WebSearchSettings: FC = () => {
-  const { providers, updateWebSearchProviders } = useWebSearchProviders()
-  console.log('providers', providers)
+  const { providers } = useWebSearchProviders()
+  const { provider: defaultProvider, setDefaultProvider } = useDefaultWebSearchProvider()
+
   const { t } = useTranslation()
-  const [selectedProvider, setSelectedProvider] = useState<WebSearchProvider | null>(null)
-  const [viewType, setViewType] = useState<'basic' | 'provider'>('basic')
+  const [selectedProvider, setSelectedProvider] = useState<WebSearchProvider | undefined>(defaultProvider)
+  const { theme: themeMode } = useTheme()
 
-  const [, setDragging] = useState(false)
+  const [apiChecking, setApiChecking] = useState(false)
+  const [apiValid, setApiValid] = useState(false)
 
-  const onDragEnd = (result: DropResult) => {
-    setDragging(false)
-    if (result.destination) {
-      const sourceIndex = result.source.index
-      const destIndex = result.destination.index
-      const reorderProviders = droppableReorder<WebSearchProvider>(providers, sourceIndex, destIndex)
-      updateWebSearchProviders(reorderProviders)
+  function updateSelectedWebSearchProvider(providerId: string) {
+    const provider = providers.find((p) => p.id === providerId)
+    if (!provider) {
+      return
+    }
+    setApiValid(false)
+    setSelectedProvider(provider)
+    setDefaultProvider(provider)
+  }
+  async function checkSearch() {
+    // 检查是否选择了提供商
+    if (!selectedProvider || !selectedProvider.id) {
+      window.message.error({
+        content: t('settings.websearch.no_provider_selected'),
+        duration: 3,
+        icon: <InfoCircleOutlined />,
+        key: 'no-provider-selected'
+      })
+      return
+    }
+
+    try {
+      setApiChecking(true)
+      const { valid, error } = await WebSearchService.checkSearch(selectedProvider)
+
+      setApiValid(valid)
+
+      // 如果验证失败且有错误信息，显示错误
+      if (!valid && error) {
+        const errorMessage = error.message ? ' ' + error.message : ''
+        window.message.error({
+          content: errorMessage,
+          duration: 4,
+          key: 'search-check-error'
+        })
+      }
+    } catch (err) {
+      console.error('Check search error:', err)
+      setApiValid(false)
+      window.message.error({
+        content: t('settings.websearch.check_failed'),
+        duration: 3,
+        key: 'check-search-error'
+      })
+    } finally {
+      setApiChecking(false)
     }
   }
 
-  const handleBasicClick = () => {
-    setViewType('basic')
-  }
-
-  const handleProviderClick = (provider: WebSearchProvider) => {
-    setSelectedProvider(provider)
-    setViewType('provider')
-  }
-
   return (
-    <Container>
-      <ProviderListContainer>
-        <Scrollbar>
-          <ProviderList>
-            <ProviderListItem
-              key={'basic'}
-              style={{ marginBottom: 5 }}
-              className={viewType === 'basic' ? 'active' : ''}
-              onClick={handleBasicClick}>
-              <SettingOutlined size={25} style={{ marginRight: '10px' }} />
-              <div style={{ fontWeight: '500', fontFamily: 'Ubuntu' }}>{t('settings.general')}</div>
-            </ProviderListItem>
-            <DragDropContext onDragStart={() => setDragging(true)} onDragEnd={onDragEnd}>
-              <Droppable droppableId="droppable">
-                {(provided) => (
-                  <div {...provided.droppableProps} ref={provided.innerRef}>
-                    {providers.map((provider, index) => (
-                      <Draggable key={`draggable_${provider.id}_${index}`} draggableId={provider.id} index={index}>
-                        {(provided) => (
-                          <div
-                            ref={provided.innerRef}
-                            {...provided.draggableProps}
-                            {...provided.dragHandleProps}
-                            style={{ ...provided.draggableProps.style, marginBottom: 5 }}>
-                            <ProviderListItem
-                              key={JSON.stringify(provider)}
-                              className={
-                                viewType === 'provider' && provider.id === selectedProvider?.id ? 'active' : ''
-                              }
-                              onClick={() => handleProviderClick(provider)}>
-                              <ProviderLogo shape="square" src={getWebSearchProviderLogo(provider.id)} size={25} />
-                              <ProviderItemName className="text-nowrap">{provider.name}</ProviderItemName>
-                              {provider.enabled && (
-                                <Tag color="green" style={{ marginLeft: 'auto', marginRight: 0, borderRadius: 16 }}>
-                                  ON
-                                </Tag>
-                              )}
-                            </ProviderListItem>
-                          </div>
-                        )}
-                      </Draggable>
-                    ))}
-                  </div>
-                )}
-              </Droppable>
-            </DragDropContext>
-          </ProviderList>
-        </Scrollbar>
-      </ProviderListContainer>
-
-      {/* 根据当前视图类型显示不同的内容 */}
-      {viewType === 'provider' && selectedProvider && (
-        <WebSearchProviderSetting provider={selectedProvider} key={JSON.stringify(selectedProvider)} />
-      )}
-      {viewType === 'basic' && <BasicSettings key={'basic'} />}
-    </Container>
+    <SettingContainer theme={themeMode}>
+      <SettingGroup theme={themeMode}>
+        <SettingTitle>{t('settings.websearch.title')}</SettingTitle>
+        <SettingDivider />
+        <SettingRow>
+          <SettingRowTitle>{t('settings.websearch.search_provider')}</SettingRowTitle>
+          <div style={{ display: 'flex', gap: '8px' }}>
+            <Select
+              value={selectedProvider?.id}
+              style={{ width: '200px' }}
+              onChange={(value: string) => updateSelectedWebSearchProvider(value)}
+              placeholder={t('settings.websearch.search_provider_placeholder')}
+              options={providers.map((p) => ({ value: p.id, label: p.name }))}
+            />
+            <Button
+              ghost={apiValid}
+              type={apiValid ? 'primary' : 'default'}
+              onClick={async () => await checkSearch()}
+              disabled={apiChecking}>
+              {apiChecking ? <LoadingOutlined spin /> : apiValid ? <CheckOutlined /> : t('settings.websearch.check')}
+            </Button>
+          </div>
+        </SettingRow>
+        <SettingDivider />
+        {selectedProvider && <WebSearchProviderSetting provider={selectedProvider} />}
+      </SettingGroup>
+      <BasicSettings />
+      <BlacklistSettings />
+    </SettingContainer>
   )
 }
-
-const Container = styled.div`
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  padding: 2px 0;
-`
-const ProviderListContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  min-width: calc(var(--settings-width) + 10px);
-  height: calc(100vh - var(--navbar-height));
-  border-right: 0.5px solid var(--color-border);
-`
-
-const ProviderList = styled.div`
-  display: flex;
-  flex: 1;
-  flex-direction: column;
-  padding: 8px;
-  padding-right: 5px;
-`
-
-const ProviderListItem = styled.div`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  padding: 5px 10px;
-  width: 100%;
-  cursor: grab;
-  border-radius: var(--list-item-border-radius);
-  font-size: 14px;
-  transition: all 0.2s ease-in-out;
-  border: 0.5px solid transparent;
-  &:hover {
-    background: var(--color-background-soft);
-  }
-  &.active {
-    background: var(--color-background-soft);
-    border: 0.5px solid var(--color-border);
-    font-weight: bold !important;
-  }
-`
-
-const ProviderLogo = styled(Avatar)`
-  border: 0.5px solid var(--color-border);
-`
-
-const ProviderItemName = styled.div`
-  margin-left: 10px;
-  font-weight: 500;
-  font-family: Ubuntu;
-`
-
 export default WebSearchSettings
